% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFetusesLiveborn.R
\name{checkFetusesLiveborn}
\alias{checkFetusesLiveborn}
\title{Title}
\usage{
checkFetusesLiveborn(worktable)
}
\arguments{
\item{worktable}{is the motherTable}
}
\value{
returns a table with the checks on fetuses and liveborns
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
checkFetusesLiveborn(cdm$mothertable)
}
}
