% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFetusId.R
\name{checkFetusId}
\alias{checkFetusId}
\title{tells us how the number of fetuses and liveborns add up in the mother table and babytable}
\usage{
checkFetusId(mothertable, babytable)
}
\arguments{
\item{mothertable}{is the mothertable}

\item{babytable}{is the babytable}
}
\value{
returns a table with the checks on fetuses and liveborns
}
\description{
tells us how the number of fetuses and liveborns add up in the mother table and babytable
}
\examples{
\donttest{
cdm <- mockPregnancy()
checkFetusId(cdm$mothertable,cdm$babytable)
}
}
