% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutcomeMode.R
\name{checkOutcomeMode}
\alias{checkOutcomeMode}
\title{Title}
\usage{
checkOutcomeMode(workTable)
}
\arguments{
\item{workTable}{is the motherTable}
}
\value{
returns a table with the information of differences in outcome mode and delivery data
}
\description{
Title
}
