% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeChecks.R
\name{executeChecks}
\alias{executeChecks}
\title{Title}
\usage{
executeChecks(
  motherTable = NULL,
  babyTable = NULL,
  checks = c("overview", "missing", "gestationalAge", "outcomeMode", "fetusesLiveborn",
    "fetusid"),
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{
all the individual result tables with data obscured if need be
}
\description{
Title
}
