% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnualOverview.R
\name{getAnnualOverview}
\alias{getAnnualOverview}
\title{Title}
\usage{
getAnnualOverview(mothertable)
}
\arguments{
\item{mothertable}{is the mothertable}
}
\value{
a table which shows the number of pregnancies per year
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
getAnnualOverview(cdm$mothertable)
}
}
