% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBitSet.R
\name{getBitSet}
\alias{getBitSet}
\title{Title}
\usage{
getBitSet(motherTable = NULL, babyTable = NULL)
}
\arguments{
\item{motherTable}{is the motherTable}

\item{babyTable}{is the babyTable}
}
\value{
returns a table with the pattern of missing data
}
\description{
Title
}
