% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissings.R
\name{getMissings}
\alias{getMissings}
\title{tells us how many missing values are there in the mother table or infant table (depending on the input)}
\usage{
getMissings(worktable)
}
\arguments{
\item{worktable}{is the mothertable or babytable}
}
\value{
returns a table with all the missings
}
\description{
tells us how many missing values are there in the mother table or infant table (depending on the input)
}
\examples{
\donttest{
cdm <- mockPregnancy()
getMissings(cdm$babytable)
}
}
