% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverview.R
\name{getOverview}
\alias{getOverview}
\title{Tells us how many women and pregnancies there are when the mother table is assessed and how many pregnancies and infants there are in case the infant table is assessed}
\usage{
getOverview(workTable)
}
\arguments{
\item{workTable}{is the motherTable or babyTable}
}
\value{
returns a table with the overview
}
\description{
Tells us how many women and pregnancies there are when the mother table is assessed and how many pregnancies and infants there are in case the infant table is assessed
}
\examples{
\donttest{
cdm <- mockPregnancy()
getOverview(cdm$babytable)
}
}
