% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnknown.R
\name{getUnknown}
\alias{getUnknown}
\title{Title}
\usage{
getUnknown(motherTable)
}
\arguments{
\item{motherTable}{is the motherTable}
}
\value{
returns a table with all the zero counts of required variables in the motherTable
}
\description{
Title
}
