% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValueDatesAgeDist.R
\name{getValueDatesAgeDist}
\alias{getValueDatesAgeDist}
\title{tells us what the minimum and maximum of pregnancy start and end dates as well as the distribution of gestational age are}
\usage{
getValueDatesAgeDist(mothertable)
}
\arguments{
\item{mothertable}{is the mothertable}
}
\value{
returns a table with the distrubtion of pregnancy starte date, end date, and gestational Age
}
\description{
tells us what the minimum and maximum of pregnancy start and end dates as well as the distribution of gestational age are
}
\examples{
\donttest{
cdm <- mockPregnancy()
getValueDatesAgeDist(cdm$mothertable)
}
}
