% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValueWeightDist.R
\name{getValueWeightDist}
\alias{getValueWeightDist}
\title{Title}
\usage{
getValueWeightDist(babyTable)
}
\arguments{
\item{babyTable}{is the babyTable}
}
\value{
returns values of birthweight
}
\description{
Title
}
