% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockDB.R
\name{mockPregnancy}
\alias{mockPregnancy}
\title{Title}
\usage{
mockPregnancy(
  mothertable = NULL,
  babytable = NULL,
  pregnancy_size = 100,
  fetus_size = 110,
  seed = 1
)
}
\arguments{
\item{mothertable}{creates a table according to the mothertable of the pregnancy extension Table in OHDSI}

\item{babytable}{creates a table according to the babytable of the pregnancy extension Table in OHDSI}

\item{pregnancy_size}{chose the number of pregnancies}

\item{fetus_size}{chose the number of fetuses, has to be equal or larger than number of pregnancies}

\item{seed}{chose a numer, like 1 for example}
}
\value{
a mock cdm reference mothertable and babytable
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
cdm
}
}
