% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETDiagnostics-package.R
\docType{package}
\name{PETDiagnostics-package}
\alias{PETDiagnostics}
\alias{PETDiagnostics-package}
\title{PETDiagnostics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Theresa Burkard \email{theresa.burkard@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-1313-4473}{ORCID})

}
\keyword{internal}
