% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseGestationalAge.R
\name{summariseGestationalAge}
\alias{summariseGestationalAge}
\title{tells us how pregnancy dates and gestational age add up in the mothertable}
\usage{
summariseGestationalAge(worktable, minGestAge_Days)
}
\arguments{
\item{worktable}{is the mothertable}

\item{minGestAge_Days}{is minimum number of days of gestational age for which we test whether the end was before (start + minGestAge_Days)}
}
\value{
returns a table with the dates checks
}
\description{
tells us how pregnancy dates and gestational age add up in the mothertable
}
\examples{
\donttest{
cdm <- mockPregnancy()
summariseGestationalAge(cdm$mothertable, minGestAge_Days = 28)
}
}
