% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeChecks.R
\name{writeResultToDisk}
\alias{writeResultToDisk}
\title{Write diagnostics results to a zip file on disk in given output folder.}
\usage{
writeResultToDisk(resultList, databaseId, outputFolder)
}
\arguments{
\item{resultList}{named list with results}

\item{databaseId}{database identifier}

\item{outputFolder}{folder to write to}
}
\description{
Write diagnostics results to a zip file on disk in given output folder.
}
