% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFetusesLiveborn.R
\name{checkFetusesLiveborn}
\alias{checkFetusesLiveborn}
\title{Title}
\usage{
checkFetusesLiveborn(workTable)
}
\arguments{
\item{workTable}{is the motherTable or babyTable}
}
\value{
returns a table with the checks on fetuses and liveborns
}
\description{
Title
}
