% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFetusId.R
\name{checkFetusId}
\alias{checkFetusId}
\title{Title}
\usage{
checkFetusId(motherTable, babyTable)
}
\arguments{
\item{motherTable}{is the motherTable}

\item{babyTable}{is the babyTable}
}
\value{
returns a tale with the fetuses checks
}
\description{
Title
}
