% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOutcomeMode.R
\name{checkOutcomeMode}
\alias{checkOutcomeMode}
\title{tells us how outcome and mode of delivery add up in the mother table}
\usage{
checkOutcomeMode(worktable)
}
\arguments{
\item{worktable}{is the mothertable}
}
\value{
returns a table with the information of differences in outcome mode and delivery data
}
\description{
tells us how outcome and mode of delivery add up in the mother table
}
\examples{
\donttest{
cdm <- mockPregnancy()
checkOutcomeMode(cdm$mothertable)
}
}
