% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeChecks.R
\name{executeChecks}
\alias{executeChecks}
\title{this function executes the desired checks "overview","annualOverview","missing", "unknown","gestationalAge","datesAgeDist","outcomeMode", "fetusesLiveborn","fetusid","weightDist","bitSet"}
\usage{
executeChecks(
  mothertable = NULL,
  babytable = NULL,
  checks = c("overview", "annualOverview", "missing", "unknown", "gestationalAge",
    "datesAgeDist", "outcomeMode", "fetusesLiveborn", "fetusid", "weightDist", "bitSet"),
  minCellCount = 5,
  minGestAge_Days = 21,
  verbose = FALSE
)
}
\arguments{
\item{mothertable}{input a table according to the mothertable of the pregnancy extension Table in OHDSI, can be NULL}

\item{babytable}{input a table according to the mothertable of the pregnancy extension Table in OHDSI, can be NULL}

\item{checks}{chose the checks you want to perform, not all checks are possible depending on which tables are available}

\item{minCellCount}{chose a number below you want to obscure counts, 0 is not obscured}

\item{minGestAge_Days}{chose a number below you think that the pregnancy entries were wrong / should not be there}

\item{verbose}{will give you the information the package is performing at that moment}
}
\value{
all the individual result tables with data obscured if need be
}
\description{
this function executes the desired checks "overview","annualOverview","missing", "unknown","gestationalAge","datesAgeDist","outcomeMode", "fetusesLiveborn","fetusid","weightDist","bitSet"
}
\examples{
\donttest{
cdm <- mockPregnancy()
executeChecks(cdm$mothertable,cdm$babytable)
}
}
