% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeChecks.R
\name{executeChecks}
\alias{executeChecks}
\title{Title}
\usage{
executeChecks(
  motherTable = NULL,
  babyTable = NULL,
  checks = c("overview", "annualOverview", "missing", "unknown", "gestationalAge",
    "datesAgeDist", "outcomeMode", "fetusesLiveborn", "fetusid", "weightDist", "bitSet"),
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{motherTable}{input a table according to the motherTable of the pregnancy extension Table in OHDSI, can be NULL}

\item{babyTable}{input a table according to the motherTable of the pregnancy extension Table in OHDSI, can be NULL}

\item{checks}{chose the checks you want to perform, not all checks are possible depending on which tables are available}

\item{minCellCount}{chose a number below you want to obscure counts, 0 is not obscured}

\item{verbose}{will give you the information the package is performing at that moment}
}
\value{
all the individual result tables with data obscured if need be
}
\description{
Title
}
