% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnualOverview.R
\name{getAnnualOverview}
\alias{getAnnualOverview}
\title{Title}
\usage{
getAnnualOverview(motherTable)
}
\arguments{
\item{motherTable}{is the motherTable}
}
\value{
a table which shows the number of pregnancies per year
}
\description{
Title
}
