% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBitSet.R
\name{getBitSet}
\alias{getBitSet}
\title{tells us how the missing/unknown data pattern of both tables, mother or baby table looks like (depending on the input)}
\usage{
getBitSet(mothertable = NULL, babytable = NULL)
}
\arguments{
\item{mothertable}{is the mothertable}

\item{babytable}{is the babytable}
}
\value{
returns a table with the pattern of missing data
}
\description{
tells us how the missing/unknown data pattern of both tables, mother or baby table looks like (depending on the input)
}
\examples{
\donttest{
cdm <- mockPregnancy()
getBitSet(cdm$mothertable,cdm$babytable)
}
}
