% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBitSet.R
\name{getBitSet}
\alias{getBitSet}
\title{Title}
\usage{
getBitSet(mothertable = NULL, babytable = NULL)
}
\arguments{
\item{mothertable}{is the mothertable}

\item{babytable}{is the babytable}
}
\value{
returns a table with the pattern of missing data
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
getBitSet(cdm$mothertable,cdm$babytable)
}
}
