% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissings.R
\name{getMissings}
\alias{getMissings}
\title{Title}
\usage{
getMissings(worktable)
}
\arguments{
\item{worktable}{is the mothertable or babytable}
}
\value{
returns a table with all the missings
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
getMissings(cdm$babytable)
}
}
