% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissings.R
\name{getMissings}
\alias{getMissings}
\title{Title}
\usage{
getMissings(workTable)
}
\arguments{
\item{workTable}{is the motherTable or babyTable}
}
\value{
returns a table with all the missings
}
\description{
Title
}
