% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverview.R
\name{getOverview}
\alias{getOverview}
\title{Title}
\usage{
getOverview(workTable)
}
\arguments{
\item{workTable}{is the motherTable or babyTable}
}
\value{
returns a table with the overview
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
getOverview(cdm$babytable)
}
}
