% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnknown.R
\name{getUnknown}
\alias{getUnknown}
\title{Title}
\usage{
getUnknown(mothertable)
}
\arguments{
\item{mothertable}{is the mothertable}
}
\value{
returns a table with all the zero counts of required variables in the mothertable
}
\description{
Title
}
\examples{
\donttest{
cdm <- mockPregnancy()
getUnknown(cdm$mothertable)
}
}
