% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnknown.R
\name{getUnknown}
\alias{getUnknown}
\title{tells us how many unknown values there are in the mother table (no unknown values possible for the infant table)}
\usage{
getUnknown(mothertable)
}
\arguments{
\item{mothertable}{is the mothertable}
}
\value{
returns a table with all the zero counts of required variables in the mothertable
}
\description{
tells us how many unknown values there are in the mother table (no unknown values possible for the infant table)
}
\examples{
\donttest{
cdm <- mockPregnancy()
getUnknown(cdm$mothertable)
}
}
