% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValueDatesAgeDist.R
\name{getValueDatesAgeDist}
\alias{getValueDatesAgeDist}
\title{Title}
\usage{
getValueDatesAgeDist(motherTable)
}
\arguments{
\item{motherTable}{is the motherTable}
}
\value{
returns a table with the distrubtion of pregnancy starte date, end date, and gestational Age
}
\description{
Title
}
