% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDB.R
\name{mockPregnancy}
\alias{mockPregnancy}
\title{Title}
\usage{
mockPregnancy(
  motherTable = NULL,
  babyTable = NULL,
  pregnancy_size = 100,
  fetus_size = 110,
  seed = 1
)
}
\arguments{
\item{seed}{}
}
\value{
a mock cdm reference motherTable and babyTable
}
\description{
Title
}
