% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDB.R
\name{mockPregnancy}
\alias{mockPregnancy}
\title{Title}
\usage{
mockPregnancy(
  motherTable = NULL,
  babyTable = NULL,
  pregnancy_size = 100,
  fetus_size = 110,
  seed = 1
)
}
\arguments{

\item{motherTable}{creates a table according to the motherTable of the pregnancy extension Table in OHDSI}

\item{babyTable}{creates a table according to the babyTable of the pregnancy extension Table in OHDSI}

\item{pregnancy_size}{chose the number of pregnancies}

\item{fetus_size}{chose the number of fetuses, has to be equal or larger than number of pregnancies}

\item{seed}{chose a numer, like 1 for example}

}
\value{
a mock cdm reference motherTable and babyTable
}
\description{
Title
}
