% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseGestationalAge.R
\name{summariseGestationalAge}
\alias{summariseGestationalAge}
\title{Title}
\usage{
summariseGestationalAge(workTable)
}
\arguments{
\item{workTable}{is the motherTable or babyTable}
}
\value{
returns a table with the dates checks
}
\description{
Title
}
